Restructure for JDA version 4 compatibility
